requirement.txt
pytorch
torchvision
torchaudio
opencv-python
pandas
tqdm
scikit-learn
scikit-image
matplotlib